requirements.txt
django>=4.2,<6
psycopg[binary]>=3.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
pymupdf>=1.24
pytesseract>=0.3.10
pillow>=10.0
sentence-transformers>=2.2
qdrant-client>=1.7
google-genai>=1.0
